+++ packages/frontend-device-intelligence/jest.config.js
@@ -1,5 +1,13 @@
 export default {
-    collectCoverage: true,
-    preset: "ts-jest",
-    testEnvironment: "jsdom",
-  };
+  collectCoverage: true,
+  collectCoverageFrom: [
+    "<rootDir>/src/**/*.ts",
+    "!<rootDir>/src/**/*.interface.ts",
+    "!<rootDir>/src/**/*.mock.ts",
+    "!<rootDir>/src/**/*.spec.ts",
+    "!<rootDir>/src/**/*.test.ts",
+    "!<rootDir>/src/**/*.d.ts",
+  ],
+  preset: "ts-jest",
+  testEnvironment: "jsdom",
+};