+++ packages/frontend-device-intelligence/CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 🚀 Features
 
 - Initialising the frontend device intelligence package
+- ThumbmarkJS cloned into the frontend device intelligence package to use code for fingerprinting.
 
 ### ❤️  Thank You